update navbar
+ fixing logo disappeared upon scrolling pass 400px Y-axis
+ logo size adjustment based on screen
+ adding logo bak file
