+++ app/components/bak/logo/Logo.tsx
@@ -1,0 +1,58 @@
+"use client";
+import Image from "next/image";
+import { useEffect, useState } from "react";
+import Link from "next/link";
+// import Button from "./Button";
+
+const Logo_old = () => {
+  //update the size of the logo when the size of the screen changes
+  const [width, setWidth] = useState(0);
+
+  const updateWidth = () => {
+    const newWidth = window.innerWidth;
+    setWidth(newWidth);
+  };
+
+  useEffect(() => {
+    window.addEventListener("resize", updateWidth);
+    updateWidth();
+  }, []);
+
+  // change between the logo and the button when the user scrolls
+  const [showButton, setShowButton] = useState(false);
+
+  const changeNavButton = () => {
+    if (window.scrollY >= 400 && window.innerWidth < 768) {
+      setShowButton(false);
+    } else {
+      setShowButton(false);
+    }
+  };
+
+  useEffect(() => {
+    window.addEventListener("scroll", changeNavButton);
+  }, []);
+
+  return (
+    <>
+      <Link href="/" style={{ display: showButton ? "none" : "block" }}>
+        <Image
+          src="/logo-forweb.svg"
+          alt="Logo"
+          width={width < 1024 ? "150" : "170"}
+          height={width < 1024 ? "45" : "74"}
+          className="relative"
+        />
+      </Link>
+      <div
+        style={{
+          display: showButton ? "block" : "none",
+        }}
+      >
+        {/* <Button /> */}
+      </div>
+    </>
+  );
+};
+
+export default Logo_old;
