+++ app/components/navigation/navbar/page.tsx
@@ -56,16 +56,14 @@ const Navbar = () => {
       <div className="flex container mx-auto px-4 py-2 justify-between items-center w-full">
         <div>
           {/* <h1 className="text-5xl font-signature ml-2"><a className="link-underline hover:transition ease-in-out delay-150 hover:underline hover:decoration-solid" href="">Logo</a></h1> */}
-          <h1 className="text-5xl font-signature ml-2">
-            <Link
-              href="/"
-              className="link-underline link-underline-black"
-              target="_blank"
-              rel="noreferrer"
-            >
-              <Logo />
-            </Link>
-          </h1>
+          <Link
+            href="/"
+            className="link-underline link-underline-black"
+            target="_blank"
+            rel="noreferrer"
+          >
+            <Logo />
+          </Link>
         </div>
 
         <ul className="hidden md:flex">
@@ -94,11 +92,11 @@ const Navbar = () => {
         </div>
 
         {nav && (
-          <ul className="flex flex-col justify-center items-center absolute top-0 left-0 w-full h-screen bg-gradient-to-b from-black to-gray-800 text-gray-500">
+          <ul className="flex flex-col justify-center items-center absolute top-0 left-0 w-full h-screen bg-gradient-to-b bg-white text-bluegreen font-medium hover:font-semibold">
             {links.map(({ id, link, name }) => (
               <li
                 key={id}
-                className="px-4 cursor-pointer capitalize py-6 text-4xl"
+                className="px-4 cursor-pointer capitalize py-6 text-xl"
               >
                 <Link onClick={() => setNav(!nav)} href={link}>
                   {name}
